% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{AIC.glmcs}
\alias{AIC.glmcs}
\title{Extract AIC from a glmcs Object}
\usage{
\method{AIC}{glmcs}(object, ..., k = 2)
}
\arguments{
\item{object}{A fitted model object of class "glmcs".}

\item{...}{Additional arguments (not used).}

\item{k}{Numeric, the penalty per parameter to be used; default is 2.}
}
\value{
AIC value.
}
\description{
Returns the Akaike Information Criterion (AIC) for a glmcs model.
}
