% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{AIC.glmsusie}
\alias{AIC.glmsusie}
\title{Extract AIC from a glmsusie Object}
\usage{
\method{AIC}{glmsusie}(object, ..., k = 2)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{...}{Additional arguments (not used).}

\item{k}{Numeric, the penalty per parameter to be used; default is 2.}
}
\value{
AIC value.
}
\description{
Returns the Akaike Information Criterion (AIC) for a glmsusie model.
}
