% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{BIC.glmcs}
\alias{BIC.glmcs}
\title{Extract BIC from a glmcs Object}
\usage{
\method{BIC}{glmcs}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmcs".}

\item{...}{Additional arguments (not used).}
}
\value{
BIC value.
}
\description{
Returns the Bayesian Information Criterion (BIC) for a glmcs model.
}
