% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{BIC.glmsusie}
\alias{BIC.glmsusie}
\title{Extract BIC from a glmsusie Object}
\usage{
\method{BIC}{glmsusie}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{...}{Additional arguments (not used).}
}
\value{
BIC value.
}
\description{
Returns the Bayesian Information Criterion (BIC) for a glmsusie model.
}
