% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{deviance.glmcs}
\alias{deviance.glmcs}
\title{Calculate Deviance for a glmcs Object}
\usage{
\method{deviance}{glmcs}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmcs".}

\item{...}{Additional arguments (not used).}
}
\value{
Deviance value.
}
\description{
Returns the deviance of a glmcs model.
}
