% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{deviance.glmsusie}
\alias{deviance.glmsusie}
\title{Calculate Deviance for a glmsusie Object}
\usage{
\method{deviance}{glmsusie}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{...}{Additional arguments (not used).}
}
\value{
Deviance value.
}
\description{
Returns the deviance of a glmsusie model.
}
