% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{dispersion.glmcs}
\alias{dispersion.glmcs}
\title{Extract Dispersion Parameter from a glmcs Object}
\usage{
dispersion.glmcs(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmcs".}

\item{...}{Additional arguments (not used).}
}
\value{
Dispersion parameter value.
}
\description{
Returns the estimated dispersion parameter from a glmcs model.
}
