% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{dispersion.glmsusie}
\alias{dispersion.glmsusie}
\title{Extract Dispersion Parameter from a glmsusie Object}
\usage{
dispersion.glmsusie(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{...}{Additional arguments (not used).}
}
\value{
Dispersion parameter value.
}
\description{
Returns the estimated dispersion parameter from a glmsusie model.
}
