% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{fitted.glmcs}
\alias{fitted.glmcs}
\title{Extract Fitted Values from a glmcs Object}
\usage{
\method{fitted}{glmcs}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmcs".}

\item{...}{Additional arguments (not used).}
}
\value{
A vector of fitted values on the scale of the response variable.
}
\description{
Returns fitted values from a glmcs model object.
}
