% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{fitted.glmsusie}
\alias{fitted.glmsusie}
\title{Extract Fitted Values from a glmsusie Object}
\usage{
\method{fitted}{glmsusie}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{...}{Additional arguments (not used).}
}
\value{
A vector of fitted values on the scale of the response variable.
}
\description{
Returns fitted values from a glmsusie model object.
}
