% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmcs-package.R
\docType{package}
\name{glmsusie-package}
\alias{glmsusie-package}
\title{glmsusie: gSuSiE: Generalized Sum of Single Effects for Generalized Linear and Cox Models}
\description{
Implements likelihood-based additive single-effect regression (LASER) via a blockwise coordinate ascent algorithm, providing confident sets for variable selection with uncertainty quantification. Features include robust variable selection under high multicollinearity, confident sets that capture posterior uncertainty in variable inclusion, flexible modeling for linear, generalized linear, and Cox regression, high performance via C++ backend, and configurable parameters for number of effects, coverage, and convergence tolerance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yizenglistat/glmsusie}
  \item Report bugs at \url{https://github.com/yizenglistat/glmsusie/issues}
}

}
\author{
\strong{Maintainer}: Yizeng Li \email{yizenglistat@gmail.com}

Authors:
\itemize{
  \item Wei Pan \email{panxx014@umn.edu}
}

}
\keyword{internal}
