% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_covered}
\alias{is_covered}
\title{Check Whether True Actives Are Covered by Confidence Sets}
\usage{
is_covered(confidence_sets, true_active)
}
\arguments{
\item{confidence_sets}{A list of integer vectors, each representing a confidence set
of selected variable indices. If a single vector is provided, it will be coerced
to a list of length one.}

\item{true_active}{Integer vector of true active variable indices.}
}
\value{
Logical scalar. \code{TRUE} if all elements of \code{true_active} are
contained in the union of \code{confidence_sets}, \code{FALSE} otherwise.
}
\description{
Determines if all true active variables are included in at least one of the
provided confidence sets. Returns \code{TRUE} if every element of \code{true_active}
appears in the union of the sets, and \code{FALSE} otherwise.
}
\examples{
# Single set, true actives 1 and 3 are both covered
is_covered(confidence_sets = c(1, 3, 5), true_active = c(1, 3))

# Multiple sets, true active 2 appears in one of them
sets <- list(cs1 = c(1,4), cs2 = c(2,5), cs3 = c(3))
is_covered(confidence_sets = sets, true_active = c(2,3))

# Not covered if any true active is missing
is_covered(confidence_sets = list(c(1,4)), true_active = c(1,2))

}
