% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{logLik.glmcs}
\alias{logLik.glmcs}
\title{Extract Log-Likelihood from a glmcs Object}
\usage{
\method{logLik}{glmcs}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmcs".}

\item{...}{Additional arguments (not used).}
}
\value{
Log-likelihood value.
}
\description{
Returns the log-likelihood of a glmcs model object.
}
