% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{logLik.glmsusie}
\alias{logLik.glmsusie}
\title{Extract Log-Likelihood from a glmsusie Object}
\usage{
\method{logLik}{glmsusie}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{...}{Additional arguments (not used).}
}
\value{
Log-likelihood value.
}
\description{
Returns the log-likelihood of a glmsusie model object.
}
