% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{plot.glmsusie}
\alias{plot.glmsusie}
\title{Plot a glmsusie Model Object}
\usage{
\method{plot}{glmsusie}(x, which = c("coefficients", "probabilities", "sets"), n_top = 20, ...)
}
\arguments{
\item{x}{A glmsusie object.}

\item{which}{Integer or character string specifying which plot to create:
\itemize{
\item 1 or "coefficients": Plot of coefficient estimates
\item 2 or "probabilities": Plot of inclusion probabilities
\item 3 or "sets": Visualization of confidence sets
}}

\item{n_top}{Integer specifying how many top coefficients to display.}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
A ggplot2 object.
}
\description{
Produces various plots for visualizing glmsusie model results.
}
