% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.glmcs}
\alias{print.glmcs}
\title{Print a glmcs Model Object}
\usage{
\method{print}{glmcs}(x, ...)
}
\arguments{
\item{x}{A glmcs object.}

\item{...}{Additional arguments passed to print.}
}
\value{
The object invisibly.
}
\description{
Print a glmcs Model Object
}
