% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.glmsusie}
\alias{print.glmsusie}
\title{Print a glmsusie Model Object}
\usage{
\method{print}{glmsusie}(x, ...)
}
\arguments{
\item{x}{A glmsusie object.}

\item{...}{Additional arguments passed to print.}
}
\value{
The object invisibly.
}
\description{
Print a glmsusie Model Object
}
