% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{residuals.glmsusie}
\alias{residuals.glmsusie}
\title{Extract Residuals from a glmsusie Object}
\usage{
\method{residuals}{glmsusie}(object, type = c("deviance", "pearson", "response", "working"), ...)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{type}{Character string indicating the type of residuals to be returned:
\itemize{
\item "deviance": deviance residuals
\item "pearson": Pearson residuals
\item "response": response residuals (observed - fitted)
\item "working": working residuals
}}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of residuals.
}
\description{
Extracts residuals from a glmsusie model fit.
}
