% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{summary.glmcs}
\alias{summary.glmcs}
\title{Summarize a glmcs Model Fit}
\usage{
\method{summary}{glmcs}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmcs".}

\item{...}{Additional arguments (not used).}
}
\value{
A list with class "summary.glmcs" containing various summary statistics.
}
\description{
Produces a summary of a glmcs model fit, including coefficients,
confidence sets, and model fit statistics.
}
