% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{summary.glmsusie}
\alias{summary.glmsusie}
\title{Summarize a glmsusie Model Fit}
\usage{
\method{summary}{glmsusie}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "glmsusie".}

\item{...}{Additional arguments (not used).}
}
\value{
A list with class "summary.glmsusie" containing various summary statistics.
}
\description{
Produces a summary of a glmsusie model fit, including coefficients,
confidence sets, and model fit statistics.
}
