% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmcs-package.R
\docType{package}
\name{glmsusie-package}
\alias{glmsusie-package}
\title{glmsusie: gSuSiE: Generalized Sum of Single Effects for Generalized Linear and Cox Models}
\description{
This package performs variable selection for regression models while accounting for uncertainty. It identifies important predictors even when they are highly correlated with each other. The method works with generalized linear models and survival analysis (Cox models). It provides credible sets that show which variables are likely to be important, along with measures of uncertainty. The package is fast thanks to optimized C++ code and offers flexible settings for different analysis needs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anonymous2012review/glmsusie}
  \item Report bugs at \url{https://github.com/anonymous2012review/glmsusie/issues}
}

}
\author{
\strong{Maintainer}: Anonymous Author \email{anonymous@example.com}

}
\keyword{internal}
